Modify DHT22 reading to only turn on pulseio when reading
Following the Adafruit library usage, each DHT22 sensor requires a dedicated,
continuously running pulseio process.  Since we're using 3 of these sensors on
the chicken-pi, that means 3 continuously running processes (utilizing 3 of the
4 cores on the Raspberry Pi 3) and a lot of heat produced.

This modification moves the DHT initialization call within the update() method
and deinits them before the end of the method.  This means that the pulseio
process is only running during the actual read of the DHT22 sensors, and not
in the "sit and wait" period between calls.

Inital testing shows a CPU temperature between 60C - 64C, down from the ~80C
when three processes were continuously running.  [Note: CPU max temp is 85C]

	modified:   display_coop_values.py