Update error handling and saving current IP in monitor_network.py
Add additional error handling to deal with observed issues in monitoring
the status of the network connection of the chicken-pi.

Add a text file for saving the current public IP address to file and reading
it back in, to minimize the number of emails sent over the gmail SMTP server.

	modified:   monitor_network.py