Fix hang in display_coop_values.py when DHT22 sensor does not respond
Previously, code patiently waited for a good read from each DHT22 sensor.  When
there is a hardware malfunction, however, the code hangs because it is patiently
waiting.  Add error catching for this case, with a timeout for tries.  If
timeout is reached, print "NO DATA" in the display screen, and move on.

	modified:   display_coop_values.py