Add error check and python <= 3.6 compatibility to chichen_graphs.py
Error check: add try..except around opening of .lonlat.txt file to check if
said file really exists, then print a helpful error message and exit gracefully
if not found.

NOTE:  Need to add such checking to all other 'hidden' file reads.

Compatibility: for python 3.6 and earlier, the datetime.strptime() function
does not allow for colons in the timezone (%z) format code.  As of present, the
timestamps provided in the NWS forecast data do include said colon.  Added code
to remove this colon.  The strptime() function is backwards compatable in 3.7+,
so no current need to check for python version and have separate code blocks to
deal with the timezone parsing.

	modified:   chicken_graphs.py